+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file. The format 
 on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). Each changelog entry gets prefixed with the category of the
 item (Added, Changed, Depreciated, Removed, Fixed, Security).
 
+## [2025.02]
+- Chore: WordPress plugin updates.
+- Update: Enabled blogGap support for group, and navigation blocks.
+- Added: Footer template and basic styles updated to the latest Figma design.
+- Added: New pattern for enhanced menu layouts that can be used with the footer and as a mega menu starter.
+- Added: New custom Tribe block for copyright text in the footer.
+
 ## [2025.01]
 - Chore: WordPress plugin updates, Composer & NPM package updates.
 - Updated: Inline link styling should now apply to all links inside the post content block that do not have a